Pull interface out of CommonKeyService & Cleanup
