+++ sdk-core/src/main/java/care/data4life/sdk/crypto/CommonKeyService.kt
@@ -19,30 +19,15 @@ package care.data4life.sdk.crypto
 import care.data4life.crypto.GCKey
 import care.data4life.crypto.KeyType
 import care.data4life.sdk.CryptoSecureStore
+import care.data4life.sdk.crypto.CryptoContract.CommonKeyService.Companion.DEFAULT_COMMON_KEY_ID
 import care.data4life.sdk.lang.D4LException
 import java.io.IOException
-
-interface CommonKeyServiceContract {
-
-    fun fetchCurrentCommonKeyId(): String?
-
-    @Throws(IOException::class)
-    fun fetchCurrentCommonKey(): GCKey?
-
-    fun fetchCommonKey(commonKeyId: String): GCKey?
-
-    fun storeCurrentCommonKeyId(commonKeyId: String)
-
-    fun storeCommonKey(commonKeyId: String, commonKey: GCKey)
-
-    fun hasCommonKey(commonKeyId: String): Boolean
-}
 
 class CommonKeyService(
         private val alias: String,
         private val storage: CryptoSecureStore,
         private val keyFactory: KeyFactory
-) : CommonKeyServiceContract {
+) : CryptoContract.CommonKeyService {
 
     override fun fetchCurrentCommonKeyId(): String {
         var commonKeyId: String? = null
@@ -89,9 +74,4 @@ class CommonKeyService(
     private fun aliasCommonKey(commonKeyId: String) = "${alias}_crypto_common_key_${commonKeyId}"
 
     private val aliasCurrentCommonKeyId = "${alias}_crypto_current_common_key_id"
-
-
-    companion object {
-        private const val DEFAULT_COMMON_KEY_ID = "00000000-0000-0000-0000-000000000000"
-    }
 }
