+++ sdk-core/src/main/java/care/data4life/sdk/crypto/CryptoContract.kt
@@ -44,7 +44,27 @@ class CryptoContract {
         operator fun contains(alias: String): Boolean
     }
 
-    interface KeyFactory {
+    internal interface CommonKeyService {
+
+        fun fetchCurrentCommonKeyId(): String?
+
+        @Throws(IOException::class)
+        fun fetchCurrentCommonKey(): GCKey?
+
+        fun fetchCommonKey(commonKeyId: String): GCKey?
+
+        fun storeCurrentCommonKeyId(commonKeyId: String)
+
+        fun storeCommonKey(commonKeyId: String, commonKey: GCKey)
+
+        fun hasCommonKey(commonKeyId: String): Boolean
+
+        companion object {
+            const val DEFAULT_COMMON_KEY_ID = "00000000-0000-0000-0000-000000000000"
+        }
+    }
+
+    internal interface KeyFactory {
         fun createGCKey(exchangeKey: ExchangeKey): GCKey
         fun createGCKeyPair(exchangeKey: ExchangeKey): GCKeyPair
     }
